Content encoding: Ensure gzip is set correctly and remove the header
We have been forcing gzip whenever it hasn't been defined for GET
requests. However, we need to set it to true also for cases where
clients provide the content-encoding header.

This is also a bug fix that removes the content-encoding header from
the response whenever gzip is set because the request library implicitly
decompresses the body, but it fails to remove the content-encoding
header, causing inconsistencies in the returned response.

In the same vein, the content-length header is removed whenever the
response is decompressed or decoded.