Include query string in HTTPError body
Currently the only facts about the original request that are included in
the logged HTTPError on request failure are the URI and method. More is
required in the case of requests POSTed to the MediaWiki API in order to
identify problematic requests. This updates the HTTPError generated on
request failure to include any query string present in the original
request. This is needed to debug persistent MW API timeouts observed in
mobileapps.